Eliminar comentarios y limpiar programas

+++ Periodo1/Unidad 1/P4_Secante.m
@@ -18,9 +18,9 @@ while error>errorPrecision
     n=n+1;
     p0=p1;
     p1=p2;
-    f0=subs(f,p0);
-    f1=subs(f,p1);
-    p2=p1-(f1*(p1-p0)/(f1-f0));
+    f0=double(subs(f,p0));
+    f1=double(subs(f,p1));
+    p2=double(p1-(f1*(p1-p0)/(f1-f0)));
     error=abs(p2-p1);
     fprintf('  n=%3.0f  p0=%9.15f  p1=%9.15f  p2=%9.15f  error=%e\n',double(n),double(p0),double(p1),double(p2),double(error));
 end
